# Docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: Backend API для управления банковскими картами
  contact:
    name: Your Name
    email: your.email@example.com
  version: "1.0"
externalDocs:
  description: README проекта
  url: https://github.com/your-repo-url
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/transfers:
    post:
      tags:
        - transfer-controller
      summary: Совершить перевод
      description: Позволяет пользователю перевести деньги между своими картами
      operationId: makeTransfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferCreateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferDTO"
  /api/cards:
    post:
      tags:
        - card-controller
      summary: Создать карту
      description: Позволяет администратору создать новую карту для пользователя
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
  /api/auth/register:
    post:
      tags:
        - auth-controller
      summary: Регистрация пользователя
      description: Регистрация нового пользователя с ролью USER
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/auth/login:
    post:
      tags:
        - auth-controller
      summary: Вход пользователя
      description: Аутентификация пользователя и получение JWT токена
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /api/auth/admin-register:
    post:
      tags:
        - auth-controller
      summary: Регистрация администратора
      description: Регистрация нового пользователя с ролью ADMIN (только для администраторов)
      operationId: registerAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/cards/{id}/status:
    patch:
      tags:
        - card-controller
      summary: Обновить статус карты
      description: "Позволяет администратору изменить статус карты (ACTIVE, BLOCKED,\
        \ EXPIRED)"
      operationId: updateStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
  /api/users:
    get:
      tags:
        - user-controller
      summary: Получить всех пользователей
      description: Возвращает список всех пользователей (только для админа)
      operationId: getAllUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
  /api/users/{username}:
    get:
      tags:
        - user-controller
      summary: Получить пользователя по имени
      description: Возвращает информацию о пользователе по его username
      operationId: getUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
  /api/transfers/card/{cardId}:
    get:
      tags:
        - transfer-controller
      summary: Получить переводы по карте
      description: Возвращает список всех переводов для указанной карты
      operationId: getCardTransfers
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferDTO"
  /api/cards/{id}:
    get:
      tags:
        - card-controller
      summary: Получить карту по ID
      description: Возвращает информацию о карте по её идентификатору
      operationId: getCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDTO"
    delete:
      tags:
        - card-controller
      summary: Удалить карту
      description: Позволяет администратору удалить карту по её идентификатору
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/cards/{id}/balance:
    get:
      tags:
        - card-controller
      summary: Посмотреть баланс карты
      description: Возвращает текущий баланс указанной карты
      operationId: getBalance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/cards/user/{userId}:
    get:
      tags:
        - card-controller
      summary: Получить карты пользователя
      description: Возвращает список карт указанного пользователя с пагинацией
      operationId: getUserCards
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardDTO"
  /api/cards/my:
    get:
      tags:
        - card-controller
      summary: Получить свои карты
      description: Возвращает список карт текущего пользователя с пагинацией
      operationId: getMyCards
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardDTO"
components:
  schemas:
    TransferCreateDTO:
      type: object
      description: DTO для создания перевода между картами
      properties:
        fromCardId:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          description: Сумма перевода
          example: 1000.5
    TransferDTO:
      type: object
      description: DTO перевода между картами
      properties:
        id:
          type: integer
          format: int64
          description: ID перевода
          example: 1
        fromCardMasked:
          type: string
          description: Номер карты отправителя (замаскированный)
          example: '**** **** **** 1234'
        toCardMasked:
          type: string
          description: Номер карты получателя (замаскированный)
          example: '**** **** **** 5678'
        amount:
          type: number
          description: Сумма перевода
          example: 1000.5
        createdAt:
          type: string
          format: date-time
          description: Дата и время создания перевода
          example: 2026-04-01T12:30:45
    CardCreateDTO:
      type: object
      description: DTO для создания новой банковской карты
      properties:
        cardNumber:
          type: string
          description: Номер карты
          example: 1234123412341234
        expirationDate:
          type: string
          format: date
          description: Дата окончания действия карты
          example: 2026-12-31
        ownerId:
          type: integer
          format: int64
          description: ID владельца карты
          example: 1
        balance:
          type: number
          description: Начальный баланс карты
          example: 1000.5
    CardDTO:
      type: object
      description: DTO для представления банковской карты
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        maskedNumber:
          type: string
          description: Маскированный номер карты
          example: '**** **** **** 1234'
        expirationDate:
          type: string
          format: date
          description: Дата окончания действия карты
          example: 2026-12-31
        status:
          type: string
          description: Статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        balance:
          type: number
          description: Баланс карты
          example: 1000.5
    RegisterRequest:
      type: object
      description: DTO для регистрации нового пользователя
      properties:
        username:
          type: string
          description: Логин пользователя
          example: john_doe
        password:
          type: string
          description: Пароль пользователя
          example: strongPassword123
    LoginRequest:
      type: object
      description: DTO для запроса логина пользователя
      properties:
        username:
          type: string
          description: Имя пользователя
          example: john_doe
        password:
          type: string
          description: Пароль пользователя
          example: P@ssw0rd
    LoginResponse:
      type: object
      description: DTO для ответа при успешном логине
      properties:
        token:
          type: string
          description: JWT токен для аутентификации
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    UserDTO:
      type: object
      description: DTO пользователя
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: john_doe
        roles:
          type: array
          description: Роли пользователя
          example:
            - ROLE_USER
            - ROLE_ADMIN
          items:
            type: string
          uniqueItems: true
    PageCardDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean